# openapi.json
{
  "components": {
    "schemas": {
...
      "BuildAdmissionOverviewRead": {
        "additionalProperties": false,
        "properties": {
          "configured_concurrency": {
            "title": "Configured Concurrency",
            "type": "integer"
          },
          "effective_concurrency": {
            "title": "Effective Concurrency",
            "type": "integer"
          },
          "entries": {
            "items": {
              "$ref": "#/components/schemas/QueuedBuildAdmissionRead"
            },
            "title": "Entries",
            "type": "array"
          },
          "waiting_count": {
            "title": "Waiting Count",
            "type": "integer"
          }
        },
        "required": [
          "configured_concurrency",
          "effective_concurrency",
          "waiting_count",
          "entries"
        ],
        "title": "BuildAdmissionOverviewRead",
        "type": "object"
      },
...
      "BuildStatus": {
        "enum": [
          "QUEUED",
          "INGESTING",
          "PARSING",
          "INDEXING",
          "ANALYZING",
          "COMPILING",
          "VALIDATING",
          "PACKAGING",
          "READY",
          "FAILED",
          "CANCELLED"
        ],
        "title": "BuildStatus",
        "type": "string"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "OperationExclusionCreate": {
        "additionalProperties": false,
        "properties": {
          "operation_key": {
            "maxLength": 160,
            "minLength": 1,
            "title": "Operation Key",
            "type": "string"
          },
          "reason": {
            "maxLength": 2000,
            "minLength": 5,
            "title": "Reason",
            "type": "string"
          }
        },
        "required": [
          "operation_key",
          "reason"
        ],
        "title": "OperationExclusionCreate",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Self-hosted API-to-MCP builder/control plane",
    "title": "MCPlica API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/build-admission": {
      "get": {
        "operationId": "get_build_admission_api_v1_build_admission_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 200,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BuildAdmissionOverviewRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Build Admission",
        "tags": [
          "builds"
        ]
      }
    },
    "/api/v1/builds": {
      "get": {
        "operationId": "list_builds_api_v1_builds_get",
        "parameters": [
          {
            "in": "query",
            "name": "project_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Project Id"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/BuildStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "page_size",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Page Size",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BuildPageRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Builds",
        "tags": [
          "builds"
        ]
      }
    },
...
    "/api/v1/builds/{build_id}": {
      "get": {
        "operationId": "get_build_api_v1_builds__build_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "build_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Build Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BuildRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Build",
        "tags": [
          "builds"
        ]
      }
    },
    "/api/v1/builds/{build_id}/ai-runs": {
      "get": {
        "operationId": "get_build_ai_runs_api_v1_builds__build_id__ai_runs_get",
        "parameters": [
          {
            "in": "path",
            "name": "build_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Build Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "page_size",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Page Size",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_BuildAIRunRead_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Build Ai Runs",
        "tags": [
          "builds"
        ]
      }
    },
    "/api/v1/builds/{build_id}/cancel": {
      "post": {
        "operationId": "cancel_build_api_v1_builds__build_id__cancel_post",
        "parameters": [
          {
            "in": "path",
            "name": "build_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Build Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BuildRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Cancel Build",
        "tags": [
          "builds"
        ]
      }
    },
    "/api/v1/builds/{build_id}/diff": {
      "get": {
        "operationId": "get_build_diff_api_v1_builds__build_id__diff_get",
        "parameters": [
          {
            "in": "path",
            "name": "build_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Build Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BuildDiffRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Build Diff",
        "tags": [
          "builds"
        ]
      }
    },
    "/api/v1/builds/{build_id}/events": {
      "get": {
        "operationId": "build_events_api_v1_builds__build_id__events_get",
        "parameters": [
          {
            "in": "path",
            "name": "build_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Build Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Build Events",
        "tags": [
          "builds"
        ]
      }
    },
    "/api/v1/builds/{build_id}/export": {
      "get": {
        "operationId": "export_build_api_v1_builds__build_id__export_get",
        "parameters": [
          {
            "in": "path",
            "name": "build_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Build Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Export Build",
        "tags": [
          "builds"
        ]
      }
    },
    "/api/v1/builds/{build_id}/manifest": {
      "get": {
        "operationId": "get_build_manifest_api_v1_builds__build_id__manifest_get",
        "parameters": [
          {
            "in": "path",
            "name": "build_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Build Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MCPManifest"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Build Manifest",
        "tags": [
          "builds"
        ]
      }
    },
    "/api/v1/builds/{build_id}/manifest/download": {
      "get": {
        "operationId": "download_build_manifest_api_v1_builds__build_id__manifest_download_get",
        "parameters": [
          {
            "in": "path",
            "name": "build_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Build Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Download Build Manifest",
        "tags": [
          "builds"
        ]
      }
    },
    "/api/v1/builds/{build_id}/operations": {
      "get": {
        "operationId": "get_build_operations_api_v1_builds__build_id__operations_get",
        "parameters": [
          {
            "in": "path",
            "name": "build_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Build Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "search",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 200,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Search"
            }
          },
          {
            "in": "query",
            "name": "method",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "GET",
                    "POST",
                    "PUT",
                    "PATCH",
                    "DELETE",
                    "HEAD",
                    "OPTIONS",
                    "TRACE"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Method"
            }
          },
          {
            "in": "query",
            "name": "scope",
            "required": false,
            "schema": {
              "default": "all",
              "enum": [
                "all",
                "current-included",
                "current-excluded",
                "build-excluded",
                "changed"
              ],
              "title": "Scope",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "page_size",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Page Size",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperationPageRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Build Operations",
        "tags": [
          "builds"
        ]
      }
    },
    "/api/v1/builds/{build_id}/validation": {
      "get": {
        "operationId": "get_build_validation_api_v1_builds__build_id__validation_get",
        "parameters": [
          {
            "in": "path",
            "name": "build_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Build Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationReportRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Build Validation",
        "tags": [
          "builds"
        ]
      }
    },
...
    "/api/v1/projects/{project_id}/operation-exclusions": {
      "get": {
        "operationId": "list_operation_exclusions_api_v1_projects__project_id__operation_exclusions_get",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/OperationExclusionRead"
                  },
                  "title": "Response List Operation Exclusions Api V1 Projects  Project Id  Operation Exclusions Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Operation Exclusions",
        "tags": [
          "builds"
        ]
      },
      "post": {
        "operationId": "create_operation_exclusion_api_v1_projects__project_id__operation_exclusions_post",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OperationExclusionCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperationExclusionRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Operation Exclusion",
        "tags": [
          "builds"
        ]
      }
    },
...
  }
}